refactor: remove ColorHelper
