+++ HimbeertoniRaidTool/HimbeertoniRaidTool.csproj
@@ -50,10 +50,9 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="ColorHelper" Version="1.8.0" />
     <PackageReference Include="DalamudPackager" Version="2.1.8" />
     <PackageReference Include="XivCommon" Version="6.0.2" />
-    <Reference Include="Mono.Cecil">
+    <Reference Include="Mono.Cecil" Condition="'$(Configuration)'=='Debug'">
       <HintPath>$(DalamudLibPath)Mono.Cecil.dll</HintPath>
       <Private>false</Private>
     </Reference>
